refactor function mapping before adding type signatures
